Major Overhaul
Library now redirects you to the ViewCard page after hitting a stack. Began laying out the Study and Browse pages. We need to implement search and the study game.
