+++ frontend/mobile/mobile/src/components/ViewCard.js
@@ -30,3 +30,5 @@ const styles = StyleSheet.create({
        // backgroundColor: '#fff'
     },
 });
+
+export default ViewCard;
